python-examples/07-lifecycle-hooks/pyproject.toml
[project]
name = "polos-lifecycle-hooks"
version = "0.1.0"
description = "Lifecycle hooks example for Polos SDK"
requires-python = ">=3.10"
dependencies = [
    "polos-sdk[openai]",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.1.0",
]
